+++ .gitattributes
@@ -1,1 +1,2 @@
 * text eol=lf
+*.lockb binary diff=lockb
