+++ index.ts
@@ -1,0 +1,3 @@
+import JSP from "./lib/JSP.ts";
+
+export default JSP;
