--- lib/index.ts
@@ -1,3 +1,0 @@
-import JSP from "./JSP";
-
-export = JSP;
