+++ lib/static/api/v1.ts
@@ -1,0 +1,6 @@
+export const api = {
+    url: 'https://jspaste.tnfangel.xyz/',
+    route: {
+        documents: 'documents/'
+    }
+}
