+++ tsconfig.json
@@ -1,39 +1,25 @@
 {
   "compilerOptions": {
     "lib": [
-      "ES2021"
+      "ESNext"
     ],
-    "module": "CommonJS",
+    "module": "esnext",
+    "target": "esnext",
+    "moduleResolution": "bundler",
     "moduleDetection": "force",
-    "target": "ES2021",
-    "strict": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "Node",
-    "declaration": true,
-    "newLine": "lf",
-    "outDir": "dist",
-    "stripInternal": true,
-    "noImplicitReturns": true,
-    "noImplicitOverride": true,
-    "noImplicitThis": true,
-    "noImplicitAny": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true,
-    "noUncheckedIndexedAccess": true,
-    "noPropertyAccessFromIndexSignature": true,
+    "allowImportingTsExtensions": true,
     "noEmit": true,
-    "useDefineForClassFields": true,
-    "isolatedModules": true,
-    "esModuleInterop": true,
-    "alwaysStrict": true,
-    "strictBindCallApply": true,
-    "strictNullChecks": true,
-    "strictFunctionTypes": true,
-    "strictPropertyInitialization": true
-  },
-  "include": [
-    "lib/**/*.ts"
-  ]
+    "composite": true,
+    "strict": true,
+    "downlevelIteration": true,
+    "skipLibCheck": true,
+    "jsx": "react-jsx",
+    "stripInternal": true,
+    "allowSyntheticDefaultImports": true,
+    "forceConsistentCasingInFileNames": true,
+    "allowJs": true,
+    "types": [
+      "bun-types"
+    ]
+  }
 }
